# source/frontend/src/app/core/api/generated/openapi-core.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/folhas": {
      "get": {
        "operationId": "PayrollController_listRuns",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Optional field name for sorting.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "direction",
            "required": false,
            "in": "query",
            "schema": {
              "default": "asc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Case-insensitive text filter for representative read endpoints.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List payroll runs."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "folha-pagamento"
        ]
      },
      "post": {
        "operationId": "PayrollController_createRun",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePayrollRunDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Create a payroll run."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "folha-pagamento"
        ]
      }
    },
    "/api/v1/folhas/{folha_id}/status": {
      "patch": {
        "operationId": "PayrollController_updateStatus",
        "parameters": [
          {
            "name": "folha_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePayrollRunStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Update payroll run status."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "folha-pagamento"
        ]
      }
    },
    "/api/v1/folhas/{folha_rescisao_id}/calcular": {
      "post": {
        "operationId": "PayrollController_calculateRun",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalculatePayrollRunDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Calculate a payroll run."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "folha-pagamento"
        ]
      }
    },
    "/api/v1/folhas/{folha_id}/massa": {
      "post": {
        "operationId": "PayrollController_populateRun",
        "parameters": [
          {
            "name": "folha_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PopulatePayrollRunDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Populate a payroll run with eligible payroll items."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "folha-pagamento"
        ]
      }
    },
    "/api/v1/folhas/{folha_id}/adiantamentos": {
      "post": {
        "operationId": "PayrollController_createAdvancePayment",
        "parameters": [
          {
            "name": "folha_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAdvancePaymentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Create and process an advance request/payment for a payroll run."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "folha-pagamento"
        ]
      }
    },
...
    "/api/v1/folha/{id}/remessa": {
      "get": {
        "operationId": "PayrollOperationsController_listRemittances",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Optional field name for sorting.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "direction",
            "required": false,
            "in": "query",
            "schema": {
              "default": "asc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Case-insensitive text filter for representative read endpoints.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List remittance requests for one payroll run."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "folha-pagamento"
        ]
      },
      "post": {
        "operationId": "PayrollOperationsController_requestRemittance",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRemittanceRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Queue a CNAB remittance generation request."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "folha-pagamento"
        ]
      }
    },
    "/api/v1/folha/{id}/retorno": {
      "post": {
        "operationId": "PayrollOperationsController_requestReturnProcessing",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProcessReturnRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Queue a CNAB return processing request for one payroll run."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "folha-pagamento"
        ]
      }
    },
    "/api/v1/gfip/gerar": {
      "post": {
        "operationId": "PayrollGfipController_requestGfipGeneration",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGfipRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Queue a GFIP/SEFIP generation request."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "folha-pagamento"
        ]
      }
    },
...
  },
  "info": {
    "title": "SGP API",
    "description": "Modern SGP NestJS API",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "Object": {
        "type": "object",
        "properties": {}
      },
...
      "CreatePayrollRunDto": {
        "type": "object",
        "properties": {
          "competenceYear": {
            "type": "number",
            "minimum": 2000,
            "maximum": 2100
          },
          "competenceMonth": {
            "type": "number",
            "minimum": 1,
            "maximum": 12
          },
          "payrollTypeId": {
            "type": "string"
          },
          "processingTypeId": {
            "type": "string"
          },
          "branchId": {
            "type": "string"
          }
        },
        "required": [
          "competenceYear",
          "competenceMonth",
          "payrollTypeId",
          "processingTypeId"
        ]
      },
      "UpdatePayrollRunStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "QUEUED",
              "PROCESSING",
              "GENERATED",
              "APPROVED",
              "PAID",
              "CANCELED",
              "FAILED",
              "CLOSED"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
      "CalculatePayrollRunDto": {
        "type": "object",
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "TOTAL",
              "RETROACTIVE"
            ],
            "default": "TOTAL"
          }
        },
        "required": [
          "mode"
        ]
      },
      "PopulatePayrollRunDto": {
        "type": "object",
        "properties": {
          "replaceCalculatedItems": {
            "type": "boolean",
            "default": true
          }
        }
      },
      "CreateAdvancePaymentDto": {
        "type": "object",
        "properties": {
          "employeeId": {
            "type": "string"
          },
          "requestedAmount": {
            "type": "string"
          },
          "approvedAmount": {
            "type": "string"
          },
          "requestedOn": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          }
        },
        "required": [
          "employeeId",
          "requestedAmount"
        ]
      },
...
      "CreateRemittanceRequestDto": {
        "type": "object",
        "properties": {
          "bankId": {
            "type": "string"
          },
          "format": {
            "type": "string",
            "default": "CNAB240"
          },
          "paymentDate": {
            "type": "string"
          },
          "launchType": {
            "type": "string",
            "default": "ACCOUNT_CREDIT"
          }
        },
        "required": [
          "bankId"
        ]
      },
      "ProcessReturnRequestDto": {
        "type": "object",
        "properties": {
          "remittanceId": {
            "type": "string"
          },
          "s3Key": {
            "type": "string"
          },
          "format": {
            "type": "string",
            "default": "CNAB240"
          },
          "returnFileName": {
            "type": "string"
          }
        },
        "required": [
          "remittanceId",
          "s3Key"
        ]
      },
      "CreateGfipRequestDto": {
        "type": "object",
        "properties": {
          "payrollRunId": {
            "type": "string"
          },
          "branchId": {
            "type": "string"
          },
          "competenceYear": {
            "type": "number",
            "minimum": 2000,
            "maximum": 2100
          },
          "competenceMonth": {
            "type": "number",
            "minimum": 1,
            "maximum": 12
          },
          "collectionCode": {
            "type": "string"
          },
          "modality": {
            "type": "string"
          }
        },
        "required": [
          "competenceYear",
          "competenceMonth",
          "collectionCode",
          "modality"
        ]
      },
...
    }
  }
}